extract ad calling logic to separate file
